# expense-authorization/src/API/EMBC.ExpenseAuthorization.Api/ETeam/LookupType.cs
namespace EMBC.ExpenseAuthorization.Api.ETeam
{
    public enum LookupType
    {
        PriorityResource = 1,
        RequestingOrgForwardTo,
        ResourceCategory,
        ResourceType,
        StatusResource,
        LeadAgencyDeptList
    }
}