# expense-authorization/src/UI/angular-eteam/src/app/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EMBC Expense Authorization",
    "contact": {
      "name": "API Team",
      "url": "https://github.com/bcgov/embc-expense-authorization",
      "email": "apiteam@example.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "v1"
  },
  "paths": {
    "/api/ExpenseAuthorization": {
      "post": {
        "tags": [
          "ExpenseAuthorization"
        ],
        "summary": "Creates an Expense Authorization Request.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Event": {
                    "type": "string",
                    "nullable": true
                  },
                  "DateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "EAFNo": {
                    "type": "string",
                    "nullable": true
                  },
                  "EMBCTaskNo": {
                    "type": "string",
                    "nullable": true
                  },
                  "RequestingOrg": {
                    "type": "string",
                    "nullable": true
                  },
                  "ResourceType": {
                    "type": "string",
                    "nullable": true
                  },
                  "AuthName": {
                    "type": "string",
                    "nullable": true
                  },
                  "AuthTelephone": {
                    "type": "string",
                    "nullable": true
                  },
                  "AuthEmail": {
                    "type": "string",
                    "nullable": true
                  },
                  "Description": {
                    "type": "string",
                    "nullable": true
                  },
                  "AmountRequested": {
                    "type": "number",
                    "format": "double",
                    "nullable": true
                  },
                  "ExpenditureNotToExceed": {
                    "type": "number",
                    "format": "double",
                    "nullable": true
                  },
                  "EocApprovals.Processing.ApprovedBy": {
                    "type": "string",
                    "nullable": true
                  },
                  "EocApprovals.Processing.Position": {
                    "type": "string",
                    "nullable": true
                  },
                  "EocApprovals.Processing.ApprovalDateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "EocApprovals.ExpenditureRequest.ApprovedBy": {
                    "type": "string",
                    "nullable": true
                  },
                  "EocApprovals.ExpenditureRequest.Position": {
                    "type": "string",
                    "nullable": true
                  },
                  "EocApprovals.ExpenditureRequest.ApprovalDateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "The optional list of files to attach to the request.",
                    "nullable": true
                  }
                }
              },
              "encoding": {
                "Event": {
                  "style": "form"
                },
                "DateTime": {
                  "style": "form"
                },
                "EAFNo": {
                  "style": "form"
                },
                "EMBCTaskNo": {
                  "style": "form"
                },
                "RequestingOrg": {
                  "style": "form"
                },
                "ResourceType": {
                  "style": "form"
                },
                "AuthName": {
                  "style": "form"
                },
                "AuthTelephone": {
                  "style": "form"
                },
                "AuthEmail": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "AmountRequested": {
                  "style": "form"
                },
                "ExpenditureNotToExceed": {
                  "style": "form"
                },
                "EocApprovals.Processing.ApprovedBy": {
                  "style": "form"
                },
                "EocApprovals.Processing.Position": {
                  "style": "form"
                },
                "EocApprovals.Processing.ApprovalDateTime": {
                  "style": "form"
                },
                "EocApprovals.ExpenditureRequest.ApprovedBy": {
                  "style": "form"
                },
                "EocApprovals.ExpenditureRequest.Position": {
                  "style": "form"
                },
                "EocApprovals.ExpenditureRequest.ApprovalDateTime": {
                  "style": "form"
                },
                "files": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Lookup/ExpenditureAuthorizationResourceTypes": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "summary": "Gets the asynchronous.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Lookup/{lookupType}": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "lookupType",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LookupType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {
          "type": "object",
          "additionalProperties": false
        }
      },
      "LookupValue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LookupType": {
        "enum": [
          "PriorityResource",
          "RequestingOrgForwardTo",
          "ResourceCategory",
          "ResourceType",
          "StatusResource",
          "UnitsOfMeasure",
          "LeadAgencyDeptList"
        ],
        "type": "string"
      }
    }
  }
}